+++ patrician/domain/day.go
@@ -1,7 +1,7 @@
 package domain
 
 // Day Seconds to change day
-const Day = float64(30)
+const Day = float32(30)
 
 // Week Seconds into week
 const Week = Day * 7
